Struct subspace_test_service::mock::MockPrimaryNode

pub struct MockPrimaryNode {
    pub task_manager: TaskManager,
    pub client: Arc<Client>,
    pub backend: Arc<Backend>,
    pub executor: NativeElseWasmExecutor<TestExecutorDispatch>,
    pub transaction_pool: Arc<FullPool<Block, Client, PrimaryChainTxPreValidator<Block, Client, FraudProofVerifier, BundleValidator<Block, Client>>>>,
    pub select_chain: FullSelectChain,
    /* private fields */
}

A mock Subspace primary node instance used for testing.

Fields
task_manager: TaskManager

TaskManager’s instance.

client: Arc<Client>

Client’s instance.

backend: Arc<Backend>

Backend.

executor: NativeElseWasmExecutor<TestExecutorDispatch>

Code executor.

transaction_pool: Arc<FullPool<Block, Client, PrimaryChainTxPreValidator<Block, Client, FraudProofVerifier, BundleValidator<Block, Client>>>>

Transaction pool.

select_chain: FullSelectChain

The SelectChain Strategy

Implementations

impl MockPrimaryNode

pub fn run_mock_primary_node( tokio_handle: Handle, key: Sr25519Keyring, base_path: BasePath ) -> MockPrimaryNode

Run a mock primary node

pub fn sync_oracle() -> Arc<dyn SyncOracle + Send + Sync>

Sync oracle for MockPrimaryNode

pub fn next_slot(&self) -> u64

Return the next slot number

pub async fn produce_slot_and_wait_for_bundle_submission( &mut self ) -> (Slot, Option<SignedOpaqueBundle<NumberFor<Block>, Hash, H256>>)

Produce a slot and wait for bundle submission

pub fn new_slot_notification_stream( &mut self ) -> TracingUnboundedReceiver<(Slot, Blake2b256Hash, Option<Sender<()>>)>

Subscribe the new slot notification

pub fn block_importing_notification_stream( &mut self ) -> TracingUnboundedReceiver<(NumberFor<Block>, Sender<()>)>

Subscribe the block importing notification

impl MockPrimaryNode

pub async fn produce_block_with_slot( &mut self, slot: Slot ) -> Result<(), Box<dyn Error>>

Produce block based on the current best block and the extrinsics in pool

pub async fn produce_block_with_extrinsics( &mut self, extrinsics: Vec<<Block as BlockT>::Extrinsic> ) -> Result<(), Box<dyn Error>>

Produce block based on the current best block and the given extrinsics

pub async fn produce_blocks(&mut self, n: u64) -> Result<(), Box<dyn Error>>

Produce n number of blocks.