subspace_test_service

Type Alias FraudProofFor

pub type FraudProofFor<Block, DomainBlock> = FraudProof<NumberFor<Block>, BlockHashFor<Block>, HeaderFor<DomainBlock>, H256>;

Helper type alias

Aliased Type

pub struct FraudProofFor<Block, DomainBlock> {
    pub domain_id: DomainId,
    pub bad_receipt_hash: <<DomainBlock as Block>::Header as Header>::Hash,
    pub maybe_mmr_proof: Option<ConsensusChainMmrLeafProof<<<Block as Block>::Header as Header>::Number, <Block as Block>::Hash, H256>>,
    pub maybe_domain_runtime_code_proof: Option<DomainRuntimeCodeAt<<<Block as Block>::Header as Header>::Number, <Block as Block>::Hash, H256>>,
    pub proof: FraudProofVariant<<<Block as Block>::Header as Header>::Number, <Block as Block>::Hash, H256, <DomainBlock as Block>::Header>,
}

Fields
domain_id: DomainIdbad_receipt_hash: <<DomainBlock as Block>::Header as Header>::Hash

Hash of the bad receipt this fraud proof targeted

maybe_mmr_proof: Option<ConsensusChainMmrLeafProof<<<Block as Block>::Header as Header>::Number, <Block as Block>::Hash, H256>>

The MMR proof for the consensus state root that is used to verify the storage proof

It is set None if the specific fraud proof variant doesn’t contain a storage proof

maybe_domain_runtime_code_proof: Option<DomainRuntimeCodeAt<<<Block as Block>::Header as Header>::Number, <Block as Block>::Hash, H256>>

The domain runtime code storage proof

It is set None if the specific fraud proof variant doesn’t require domain runtime code or the required domain runtime code is available from the current runtime state.

proof: FraudProofVariant<<<Block as Block>::Header as Header>::Number, <Block as Block>::Hash, H256, <DomainBlock as Block>::Header>

The specific fraud proof variant