Crate subspace_test_service

Subspace test service only.

Macros

produce_block_with

Producing one block for both the primary node and the domain nodes, where the primary node can use the produce_block_with_xxx function (i.e. produce_block_with_slot) to produce block

produce_blocks

Produce the given number of blocks for both the primary node and the domain nodes

produce_blocks_until

Keep producing block with a fixed interval until the given condition become true

Structs

MockConsensusNode

A mock Subspace consensus node instance used for testing.

MockConsensusNodeRpcConfig

Configuration values required to run a mock consensus node with custom RPC options.

Functions

construct_extrinsic_generic

Construct an extrinsic that can be applied to the test runtime.

node_config

Create a Subspace Configuration.

Type Aliases

FraudProofFor

Helper type alias